refactor to typescript project
