+++ core/pyright-language-server/src/pythonLanguageServerConfig.json
@@ -1,0 +1,4 @@
+{
+  "languageServerDir": "../node_modules/pyright/dist",
+  "amberConfigFilePath": "../../amber/src/main/resources/application.conf"
+}
